Rename NewTable to FilterableTable
