+++ client/app/hearingSchedule/components/AssignHearingsTabs.jsx
@@ -5,7 +5,7 @@ import moment from 'moment';
 import _ from 'lodash';
 import LEGACY_APPEAL_TYPES_BY_ID from '../../../constants/LEGACY_APPEAL_TYPES_BY_ID.json';
 
-import NewTable from '../../components/NewTable';
+import FilterableTable from '../../components/FilterableTable';
 import TabWindow from '../../components/TabWindow';
 import DocketTypeBadge from '../../components/DocketTypeBadge';
 import { renderAppealType } from '../../queue/utils';
@@ -52,7 +52,7 @@ const AvailableVeteransTable = ({ rows, columns }) => {
     </div>;
   }
 
-  return <NewTable
+  return <FilterableTable
     columns={removeTimeColumn}
     rowObjects={rows}
     summary="scheduled-hearings-table"
@@ -65,7 +65,7 @@ const UpcomingHearingsTable = ({ rows, columns, selectedHearingDay }) => (
   <div>
     <Link to={`/schedule/docket/${selectedHearingDay.id}`}>
       {`View the Daily Docket for ${moment(selectedHearingDay.scheduledFor).format('M/DD/YYYY')}` }</Link>
-    <NewTable
+    <FilterableTable
       columns={columns}
       rowObjects={rows}
       summary="scheduled-hearings-table"
@@ -159,7 +159,6 @@ export default class AssignHearingsTabs extends React.Component {
   }
 
   availableVeteransRows = (appeals) => {
-
     /*
       Sorting by docket number within each category of appeal:
       CAVC, AOD and normal. Prepended * and + to docket number for
